Update fitting ROIs on tab switch
And update ROI on thumbnail (too entwined to separate)
