+++ mantidimaging/gui/windows/spectrum_viewer/fitting_form.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from logging import getLogger
 
+import numpy as np
 from PyQt5.QtWidgets import QWidget, QVBoxLayout
 
 from mantidimaging.gui.widgets.spectrum_widgets.fitting_selection_widget import FitSelectionWidget
@@ -20,10 +21,10 @@ class FittingParamFormWidgetView(QWidget):
 
     def __init__(self, spectrum_viewer: SpectrumViewerWindowView):
         super().__init__(None)
+        self.spectrum_viewer = spectrum_viewer
         self.presenter = FittingParamFormWidgetPresenter(self)
         self.setLayout(QVBoxLayout())
 
-        self.spectrum_viewer = spectrum_viewer
         self.roiSelectionWidget = ROISelectionWidget(self)
         self.layout().addWidget(self.roiSelectionWidget)
 
@@ -32,6 +33,10 @@ class FittingParamFormWidgetView(QWidget):
 
         self.fitting_param_form = FittingParamFormWidget(self.spectrum_viewer.presenter)
         self.layout().addWidget(self.fitting_param_form)
+
+        self.fittingDisplayWidget = spectrum_viewer.fittingDisplayWidget
+
+        self.roiSelectionWidget.selectionChanged.connect(self.presenter.update_roi_on_fitting_thumbnail)
 
     def showEvent(self, ev) -> None:
         super().showEvent(ev)
@@ -42,6 +47,30 @@ class FittingParamFormWidgetPresenter:
 
     def __init__(self, view: FittingParamFormWidgetView):
         self.view = view
+        self.spectrum_viewer = view.spectrum_viewer
+
+    @property
+    def fitting_spectrum(self) -> np.ndarray:
+        selected_fitting_roi = self.view.roiSelectionWidget.current_roi_name
+        if (spectrum_data := self.spectrum_viewer.spectrum_widget.spectrum_data_dict[selected_fitting_roi]) is not None:
+            return spectrum_data
+
+        raise RuntimeError("Fitting spectrum not calculated")
 
     def handle_activated(self) -> None:
         LOG.warning("Fitting form activated")
+        self.update_roi_dropdown()
+        self.update_roi_on_fitting_thumbnail()
+        self.set_default_fitting_region()
+
+    def update_roi_dropdown(self) -> None:
+        roi_names = self.spectrum_viewer.presenter.get_roi_names()
+        self.view.roiSelectionWidget.update_roi_list(roi_names)
+
+    def update_roi_on_fitting_thumbnail(self) -> None:
+        roi_widget = self.spectrum_viewer.spectrum_widget.roi_dict[self.view.roiSelectionWidget.current_roi_name]
+        self.spectrum_viewer.fittingDisplayWidget.show_roi_on_thumbnail_from_widget(roi_widget)
+
+    def set_default_fitting_region(self) -> None:
+        self.view.fittingDisplayWidget.set_default_region_if_needed(self.spectrum_viewer.presenter.model.tof_data,
+                                                                    self.fitting_spectrum)
